Edited Lagersystem and started on Mvc
New

+++ H1-Projects/Lagersystem_project/Lagersystem_project/Program.cs
@@ -12,27 +12,17 @@ namespace Lagersystem_project
             SqlConnection conn = db.connection();
             conn.Open();
 
-            bool start = false;
             const ConsoleKey keyInfo1 = ConsoleKey.D1;
             const ConsoleKey keyInfo2 = ConsoleKey.D2;
             const ConsoleKey keyInfo3 = ConsoleKey.D3;
             const ConsoleKey keyInfo4 = ConsoleKey.D4;
             const ConsoleKey keyInfo5 = ConsoleKey.D5;
+            const ConsoleKey keyInfo6 = ConsoleKey.Escape;
 
             while (true)
             {
-                if (start == true)
-                {
-                    Console.WriteLine();
-                }
-                start = true;
 
-                Console.WriteLine("Menu:");
-                Console.WriteLine("Press 1 Insert product.\n" +
-                                  "Press 2 Delete product.\n" +
-                                  "Press 3 Update product.\n" +
-                                  "Press 4 Show all product.\n" +
-                                  "Press 5 Show all locations of products\n");
+                Methods_And_Functions.Menu();
 
                 ConsoleKey pressedKey = PressedKey();
 
@@ -41,26 +31,42 @@ namespace Lagersystem_project
                     case keyInfo1:
                         Console.Write("Enter Product name > ");
                         string inputProduct = Console.ReadLine();
-                        Console.Write("Enter Product Amount > ");
-                        int productAmount = int.Parse(Console.ReadLine());
-                        Methods_And_Functions.InsertProduct(inputProduct, productAmount, conn);
-                        break;
+                        if(inputProduct != "")
+                        {
+                            Console.Write("Enter Product Amount > ");
+                            int productAmount = int.Parse(Console.ReadLine());
+                            Methods_And_Functions.InsertProduct(inputProduct, productAmount, conn);
+                            break;
+                        }
+                        Console.Clear();
+                        break;  
 
                     case keyInfo2:
                         Console.Write("Specify what ProductID you want to delete > ");
                         int productID = int.Parse(Console.ReadLine());
-                        Methods_And_Functions.DeleteProduct(productID, conn);
+                        if(productID != null)
+                        {
+                            Methods_And_Functions.DeleteProduct(productID, conn);
+                            break;
+                        }
+                        Console.Clear();
                         break;
 
                     case keyInfo3:
                         Console.Write("What ID do you want to update > ");
                         int updateID = int.Parse(Console.ReadLine());
-                        Console.Write("What do you want to update your product to > ");
-                        string updateProductName = Console.ReadLine();
-                        Console.Write("The amount of the specific product > ");
-                        int updateProductAmount = int.Parse(Console.ReadLine());
-                        Methods_And_Functions.UpdateProduct(updateID, updateProductName, updateProductAmount, conn);
+                        if( updateID != null)
+                        {
+                            Console.Write("What do you want to update your product to > ");
+                            string updateProductName = Console.ReadLine();
+                            Console.Write("The amount of the specific product > ");
+                            int updateProductAmount = int.Parse(Console.ReadLine());
+                            Methods_And_Functions.UpdateProduct(updateID, updateProductName, updateProductAmount, conn);
+                            break;
+                        }
+                        Console.Clear();
                         break;
+                        
 
                     case keyInfo4:
                         Console.WriteLine("");
@@ -85,7 +91,7 @@ namespace Lagersystem_project
                 {
                     while (!Console.KeyAvailable) ;
                     ConsoleKey pressed = Console.ReadKey(true).Key;
-                    if (pressed == keyInfo1 || pressed == keyInfo2 || pressed == keyInfo3 || pressed == keyInfo4 || pressed == keyInfo5)
+                    if (pressed == keyInfo1 || pressed == keyInfo2 || pressed == keyInfo3 || pressed == keyInfo4 || pressed == keyInfo5 || pressed == keyInfo6)
                         return pressed;
                     Console.Clear();
                 } while (true);
